LayoutTests/http/tests/security/originHeader/origin-header-for-https.html

<html>
<head>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
}
window.location = "https://127.0.0.1:8443/security/originHeader/resources/origin-header-post-to-http.html";
</script>
</head>
<body>
</body>
</html>
